requirements/base.txt
# django-ulid/requirements/base.txt
#
# Requirements for using this package.

-r build.txt

typing==3.7.4.1; python_version < '3.6'

python-ulid[pydantic]==2.6.0
